refactor : 책임 변경
