+++ src/main/java/laddergame/model/Ladder.java
@@ -8,15 +8,8 @@ import java.util.stream.Stream;
 public class Ladder {
 
   private final List<Line> lines;
-  LadderPositions resultPositions;
-
   public Ladder(List<Line> lines) {
     this.lines = lines;
-    this.resultPositions = new LadderPositions(
-        Stream.iterate(0, n -> n + 1)
-            .limit(lines.get(0).getPoints().size())
-            .map(n -> getResultPosition(lines, new Position(new NaturalNumber(n))))
-            .collect(Collectors.toList()));
   }
 
   public static Ladder createByHeightAndCountOfPerson(
@@ -30,20 +23,8 @@ public class Ladder {
 
   }
 
-  private Position getResultPosition(List<Line> lines, Position startPosition) {
-    Position position = new Position(new NaturalNumber(startPosition.getValue()));
-
-    lines.forEach(line -> position.movePositions(position, line));
-
-    return position;
-  }
-
   public List<Line> getLines() {
     return Collections.unmodifiableList(lines);
-  }
-
-  public LadderPositions getResultPositions() {
-    return resultPositions;
   }
 
 }
