+++ src/main/java/laddergame/model/LadderResultPositions.java
@@ -7,12 +7,13 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import java.util.stream.Stream;
 
-public class LadderPositions {
+public class LadderResultPositions {
 
   private final List<Position> positions;
 
-  public LadderPositions(List<Position> positions) {
+  public LadderResultPositions(List<Position> positions) {
     if (new HashSet<>(positions).size() != positions.size()) {
       throw new IllegalArgumentException("사다리 위치는 중복될 수 없습니다. positions : " + positions);
     }
@@ -20,8 +21,21 @@ public class LadderPositions {
     this.positions = positions;
   }
 
-  public static LadderPositions createResultPositionsByLadderWithCountOfNumber() {
+  public static LadderResultPositions createByLadderWithCountOfPerson(Ladder ladder,
+      PositiveNumber countOfPerson) {
+    return new LadderResultPositions(
+        Stream.iterate(0, n -> n + 1)
+            .limit(countOfPerson.getValue())
+            .map(n -> getResultPosition(ladder.getLines(), new Position(new NaturalNumber(n))))
+            .collect(Collectors.toList()));
+  }
 
+  private static Position getResultPosition(List<Line> lines, Position startPosition) {
+    Position position = new Position(new NaturalNumber(startPosition.getValue()));
+
+    lines.forEach(line -> position.movePositions(position, line));
+
+    return position;
   }
 
   public Map<Player, String> getMapOf(Players players, String[] results) {
@@ -40,7 +54,7 @@ public class LadderPositions {
     if (o == null || getClass() != o.getClass()) {
       return false;
     }
-    LadderPositions that = (LadderPositions) o;
+    LadderResultPositions that = (LadderResultPositions) o;
     return positions.equals(that.positions);
   }
